Previous/requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
Pillow
PyYAML
tqdm
